chore: ajustes laravel pint e ajustado duplicidade nos códigos de notificação de treino
