+++ app/Providers/EventServiceProvider.php
@@ -2,14 +2,14 @@
 
 namespace App\Providers;
 
+use App\Models\TeamsUsers;
 use App\Models\Training;
+use App\Observers\TeamsUsersObserver;
 use App\Observers\TrainingObserver;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Event;
-use App\Models\TeamsUsers;
-use App\Observers\TeamsUsersObserver;
 
 class EventServiceProvider extends ServiceProvider
 {
