+++ tests/Feature/GraphQL/TrainingTest.php
@@ -5,7 +5,6 @@ namespace Tests\Feature\GraphQL;
 use App\Models\Team;
 use App\Models\TeamsUsers;
 use App\Models\Training;
-use App\Models\User;
 use Faker\Factory as Faker;
 use Tests\TestCase;
